refactor(import-export): rename capability.ts and ports.ts (#1191)
- capability.ts → capabilityUtils.ts (P2: functions-only file)
- ports.ts → types.ts (P2: types-only aggregation file)

Rollback: git revert <hash> --no-edit
